add celerio conf so ave-cesar can play with it remove some hard coded package fix test compilation and make it pass

+++ src/main/sql/H2/01-create.sql
@@ -104,7 +104,7 @@ CREATE TABLE BOOK_STORE (
 );
 
 CREATE TABLE SCR (	
-	SCR_IDE int not null, 
+	SCR_IDE int8 not null,
 	SCR_CAT VARCHAR(50), 
 	SCR_TXT CLOB, 
 	SCR_AVL int,
